+++ packages/insomnia/src/ui/components/workspace-page-header.tsx
@@ -7,7 +7,8 @@ import { initializeSpectral, isLintError } from '../../common/spectral';
 import * as models from '../../models';
 import { isRemoteProject } from '../../models/project';
 import { isCollection, isDesign } from '../../models/workspace';
-import { useGitVCS, useVCS } from '../containers/app';
+import { useGitVCS } from '../hooks/use-git-vcs';
+import { useVCS } from '../hooks/use-vcs';
 import { setActiveActivity } from '../redux/modules/global';
 import { selectActiveActivity, selectActiveApiSpec, selectActiveGitRepository, selectActiveProject, selectActiveProjectName, selectActiveWorkspace, selectIsLoggedIn } from '../redux/selectors';
 import { ActivityToggle } from './activity-toggle/activity-toggle';
